+++ src/score.rs
@@ -155,3 +155,333 @@ impl Score {
         colors
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::str::FromStr;
+
+    use indexmap::IndexMap;
+
+    use crate::color::Argb;
+    use crate::Error;
+
+    use super::Score;
+
+    #[test]
+    fn test_prioritizes_chroma() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("000000")?, 1),
+            (Argb::from_str("ffffff")?, 1),
+            (Argb::from_str("0000ff")?, 1),
+        ]);
+
+        let ranked = Score::score(&argb_to_population, None, None, None);
+
+        assert_eq!(ranked.len(), 1);
+        assert_eq!(ranked[0], Argb::from_str("0000ff")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_prioritizes_chroma_when_proportions_equal() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("ff0000")?, 1),
+            (Argb::from_str("00ff00")?, 1),
+            (Argb::from_str("0000ff")?, 1),
+        ]);
+
+        let ranked = Score::score(&argb_to_population, None, None, None);
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("ff0000")?);
+        assert_eq!(ranked[1], Argb::from_str("00ff00")?);
+        assert_eq!(ranked[2], Argb::from_str("0000ff")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generates_gblue_when_no_colors_available() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> =
+            IndexMap::from_iter([(Argb::from_str("000000")?, 1)]);
+
+        let ranked = Score::score(&argb_to_population, None, None, None);
+
+        assert_eq!(ranked.len(), 1);
+        assert_eq!(ranked[0], Argb::from_str("4285f4")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_dedupes_nearby_hues() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("008772")?, 1),
+            (Argb::from_str("318477")?, 1),
+        ]);
+
+        let ranked = Score::score(&argb_to_population, None, None, None);
+
+        assert_eq!(ranked.len(), 1);
+        assert_eq!(ranked[0], Argb::from_str("008772")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_maximizes_hue_distance() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("008772")?, 1),
+            (Argb::from_str("008587")?, 1),
+            (Argb::from_str("007ebc")?, 1),
+        ]);
+
+        let ranked = Score::score(&argb_to_population, Some(2), None, None);
+
+        assert_eq!(ranked.len(), 2);
+        assert_eq!(ranked[0], Argb::from_str("007ebc")?);
+        assert_eq!(ranked[1], Argb::from_str("008772")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_one() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("7ea16d")?, 67),
+            (Argb::from_str("d8ccae")?, 67),
+            (Argb::from_str("835c0d")?, 49),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(3),
+            Some(Argb::from_str("8d3819")?),
+            Some(false),
+        );
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("7ea16d")?);
+        assert_eq!(ranked[1], Argb::from_str("d8ccae")?);
+        assert_eq!(ranked[2], Argb::from_str("835c0d")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_two() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("d33881")?, 14),
+            (Argb::from_str("3205cc")?, 77),
+            (Argb::from_str("0b48cf")?, 36),
+            (Argb::from_str("a08f5d")?, 81),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            None,
+            Some(Argb::from_str("7d772b")?),
+            None,
+        );
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("3205cc")?);
+        assert_eq!(ranked[1], Argb::from_str("a08f5d")?);
+        assert_eq!(ranked[2], Argb::from_str("d33881")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_three() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("be94a6")?, 23),
+            (Argb::from_str("c33fd7")?, 42),
+            (Argb::from_str("899f36")?, 90),
+            (Argb::from_str("94c574")?, 82),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(3),
+            Some(Argb::from_str("aa79a4")?),
+            None,
+        );
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("94c574")?);
+        assert_eq!(ranked[1], Argb::from_str("c33fd7")?);
+        assert_eq!(ranked[2], Argb::from_str("be94a6")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_four() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("df241c")?, 85),
+            (Argb::from_str("685859")?, 44),
+            (Argb::from_str("d06d5f")?, 34),
+            (Argb::from_str("561c54")?, 27),
+            (Argb::from_str("713090")?, 88),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(5),
+            Some(Argb::from_str("58c19c")?),
+            Some(false),
+        );
+
+        assert_eq!(ranked.len(), 2);
+        assert_eq!(ranked[0], Argb::from_str("df241c")?);
+        assert_eq!(ranked[1], Argb::from_str("561c54")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_five() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("be66f8")?, 41),
+            (Argb::from_str("4bbda9")?, 88),
+            (Argb::from_str("80f6f9")?, 44),
+            (Argb::from_str("ab8017")?, 43),
+            (Argb::from_str("e89307")?, 65),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(3),
+            Some(Argb::from_str("916691")?),
+            Some(false),
+        );
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("ab8017")?);
+        assert_eq!(ranked[1], Argb::from_str("4bbda9")?);
+        assert_eq!(ranked[2], Argb::from_str("be66f8")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_six() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("18ea8f")?, 93),
+            (Argb::from_str("327593")?, 18),
+            (Argb::from_str("066a18")?, 74),
+            (Argb::from_str("fa8a23")?, 62),
+            (Argb::from_str("04ca1f")?, 65),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(2),
+            Some(Argb::from_str("4c377a")?),
+            Some(false),
+        );
+
+        assert_eq!(ranked.len(), 2);
+        assert_eq!(ranked[0], Argb::from_str("18ea8f")?);
+        assert_eq!(ranked[1], Argb::from_str("fa8a23")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_seven() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("2e05ed")?, 23),
+            (Argb::from_str("153e55")?, 90),
+            (Argb::from_str("9ab220")?, 23),
+            (Argb::from_str("153379")?, 66),
+            (Argb::from_str("68bcc3")?, 81),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(2),
+            Some(Argb::from_str("f588dc")?),
+            None,
+        );
+
+        assert_eq!(ranked.len(), 2);
+        assert_eq!(ranked[0], Argb::from_str("2e05ed")?);
+        assert_eq!(ranked[1], Argb::from_str("9ab220")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_eight() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("816ec5")?, 24),
+            (Argb::from_str("6dcb94")?, 19),
+            (Argb::from_str("3cae91")?, 98),
+            (Argb::from_str("5b542f")?, 25),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(1),
+            Some(Argb::from_str("84b0fd")?),
+            Some(false),
+        );
+
+        assert_eq!(ranked.len(), 1);
+        assert_eq!(ranked[0], Argb::from_str("3cae91")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_nine() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("206f86")?, 52),
+            (Argb::from_str("4a620d")?, 96),
+            (Argb::from_str("f51401")?, 85),
+            (Argb::from_str("2b8ebf")?, 3),
+            (Argb::from_str("277766")?, 59),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            Some(3),
+            Some(Argb::from_str("02b415")?),
+            None,
+        );
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("f51401")?);
+        assert_eq!(ranked[1], Argb::from_str("4a620d")?);
+        assert_eq!(ranked[2], Argb::from_str("2b8ebf")?);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_generated_scenario_ten() -> Result<(), Error> {
+        let argb_to_population: IndexMap<Argb, u32> = IndexMap::from_iter([
+            (Argb::from_str("8b1d99")?, 54),
+            (Argb::from_str("27effe")?, 43),
+            (Argb::from_str("6f558d")?, 2),
+            (Argb::from_str("77fdf2")?, 78),
+        ]);
+
+        let ranked = Score::score(
+            &argb_to_population,
+            None,
+            Some(Argb::from_str("5e7a10")?),
+            None,
+        );
+
+        assert_eq!(ranked.len(), 3);
+        assert_eq!(ranked[0], Argb::from_str("27effe")?);
+        assert_eq!(ranked[1], Argb::from_str("8b1d99")?);
+        assert_eq!(ranked[2], Argb::from_str("6f558d")?);
+
+        Ok(())
+    }
+}