+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **added:** Add `CHANGELOG`
 - **added:** Add documentation above `struct:Argb`
 - **added:** Add `check` job to `actions:CI.yml`
+- **added:** Implement tests from C++ for `mod:scheme`, `mod:score` and `mod:temperature`.
 - **changed:** Rename `actions:tests.yml` to `actions:CI.yml`
 - **changed:** Rename `build` job to `test` in `actions:CI.yml`
 - **changed:** Remove unnecessary comments about warnings count from the `README`